+++ lib/super/error.rb
@@ -24,9 +24,9 @@ module Super
         )
       end
     end
-    # Error raised when something wasn't initalized correctly, and if there isn't
+    # Error raised when something wasn't initialized correctly, and if there isn't
     # a more specific error
-    class Initalization < Error; end
+    class Initialization < Error; end
     class ArgumentError < Error; end
     class AlreadyRegistered < Error; end
     class AlreadyTranscribed < Error; end